Remove old note about 32-bit size. See #53.

+++ inc/BitFunnel/Index/DocumentHandle.h
@@ -33,27 +33,27 @@ namespace BitFunnel
     class Slice;
     class Term;
 
-    // BITFUNNELTYPES
-    // A unique, system-wide document identifier.
-    // For now we will make not assumptions about how DocIDs are assigned to
-    // documents. The mapping may have significant and seemingly random gaps.
-    // DESIGN NOTE: The concept of invalid documents (documents with invalid
-    // document ID) was introduced to allow the ResultsProcessor to pass over
-    // document positions used to pad the DocTable row length quanta. It also
-    // provides a means to invalidate a document position after index
-    // construction.
+    // BITFUNNELTYPES A unique, system-wide document identifier.  For now we
+    // will make not assumptions about how DocIDs are assigned to documents. The
+    // mapping may have significant and seemingly random gaps.  DESIGN NOTE: The
+    // concept of invalid documents (documents with invalid document ID) was
+    // introduced to allow the ResultsProcessor to pass over document positions
+    // used to pad the DocTable row length quanta. It also provides a means to
+    // invalidate a document position after index construction.
     typedef size_t DocId;
 
 
-    // BITFUNNELTYPES
-    // A shard-independent document identifier which is local to a BitFunnel index.
-    // DocIndex values in an index run from 0 to n where n-1 is the number of documents
-    // in the index. The number of bits for DocIndex and ShardId together must not
-    // exceed 32.
+    // BITFUNNELTYPES A shard-independent document identifier which is local to
+    // a BitFunnel index.  DocIndex values in an index run from 0 to n where n-1
+    // is the number of documents in the index. See
+    // https://github.com/BitFunnel/BitFunnel/issues/53 for a potential issue
+    // about the size of DocIndex and DocId.
     typedef size_t DocIndex;
 
-    // Represent the value that the default constructor assigns to the instances of DocumentHandle.
-    static const DocIndex c_invalidDocIndex = std::numeric_limits<uint32_t>::max();
+    // Represent the value that the default constructor assigns to the instances
+    // of DocumentHandle.
+    static const DocIndex c_invalidDocIndex =
+        std::numeric_limits<size_t>::max();
 
 
     //*************************************************************************
